+++ apps/orders/models.py
@@ -31,25 +31,13 @@ class Order(BaseModel):
 
     def get_total_cost(self) -> int:
         return sum(item.get_cost() for item in self.items.all())
-    
-    def get_stripe_url(self):
-        if not self.stripe_id:
-            return
-        if '_test_' in settings.STRIPE_SECRET_KEY:
-            path = '/test/'
-        else:
-            path = '/'
-        return f"https://dashboard.stripe.com{path}payments/{self.stripe_id}"
 
     def get_stripe_url(self):
         if not self.stripe_id:
-            # No payment associated
-            return ''
+            return ""
         if '_test_' in settings.STRIPE_SECRET_KEY:
-            # For test payments
             path = '/test/'
         else:
-            # For real payments
             path = '/'
         return f"https://dashboard.stripe.com{path}payments/{self.stripe_id}"
