+++ apps/shop/models.py
@@ -1,11 +1,12 @@
 from django.db import models
 from django.urls import reverse
 from apps.common.models import BaseModel
+from django.core.validators import MinValueValidator
 
 
 class Category(BaseModel):
-    name = models.CharField(max_length=200)
-    slug = models.SlugField(max_length=200, unique=True)
+    name = models.CharField(max_length=150, null=False)
+    slug = models.SlugField(max_length=150, unique=True)
 
     class Meta:
         ordering = ["name"]
@@ -27,13 +28,15 @@ class Product(BaseModel):
         Category, related_name="products", on_delete=models.CASCADE
     )
     name = models.CharField(max_length=200)
-    slug = models.SlugField(max_length=200)
+    slug = models.SlugField(max_length=200, unique=True)
     image = models.ImageField(
         upload_to="products/%Y/%m/%d", blank=True
     )
     description = models.TextField(blank=True)
     # Avoid FloatField due to rounding issues.
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.DecimalField(max_digits=10, decimal_places=2, validators=[
+        MinValueValidator(1)
+    ])
     available = models.BooleanField(default=True)
 
     class Meta:
