+++ tox.ini
@@ -1,6 +1,15 @@
 [tox]
 envlist = py37,py38,py39,py310
 
+[pytest]
+DJANGO_SETTINGS_MODULE = config.test_settings
+python_files = test_*.py
+
+console_output_style = progress
+addopts = 
+    -n 8
+    --disable-warnings
+
 [testenv]
 deps = -r{toxinidir}/requirements-dev.txt
 commands =
